Enhance clinic and consultorio user linkage: Update saveEdit methods in Clinica and Consultorio models to automatically link clinics and consultorios to users upon creation or editing. Introduce is_new parameter to saveEdit methods in ClinicaUser and ConsultorioUser models to manage existing configurations more effectively. Improve error handling in payment registration process in principal.js for better user feedback.

+++ app/Models/ClinicaUser.php
@@ -36,11 +36,11 @@ class ClinicaUser extends Model
         return $my_clinics;
     }
 
-    public static function saveEdit($id, $request)
+    public static function saveEdit($id, $request, $is_new = false)
     {
         $existingConfiguration = ClinicaUser::where(['user_id' => $id]);
 
-        if ($existingConfiguration != null) {
+        if ($existingConfiguration != null && $is_new == false) {
             $existingConfiguration->delete();
         }
